# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AttendanceOut": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "person_name": {
            "title": "Person Name",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AttendanceType"
          }
        },
        "required": [
          "id",
          "person_id",
          "person_name",
          "type",
          "confidence",
          "created_at"
        ],
        "title": "AttendanceOut",
        "type": "object"
      },
      "AttendancePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttendanceOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "AttendancePage",
        "type": "object"
      },
      "AttendanceType": {
        "enum": [
          "ENTRY",
          "EXIT"
        ],
        "title": "AttendanceType",
        "type": "string"
      },
...
      "ConfigOut": {
        "additionalProperties": false,
        "properties": {
          "attendance_min_interval_minutes": {
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Attendance Min Interval Minutes",
            "type": "integer"
          },
          "camera_fps": {
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Camera Fps",
            "type": "integer"
          },
          "camera_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.\\-]+$",
            "title": "Camera Id",
            "type": "string"
          },
          "face_recognition_threshold": {
            "maximum": 1.0,
            "minimum": 0.2,
            "title": "Face Recognition Threshold",
            "type": "number"
          },
          "max_faces": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Faces",
            "type": "integer"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacyOut",
            "default": {
              "embeddings_encrypted": true,
              "local_processing": true,
              "save_images": false,
              "save_video": false
            }
          },
          "recognition_cooldown_seconds": {
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Recognition Cooldown Seconds",
            "type": "integer"
          },
          "save_events": {
            "title": "Save Events",
            "type": "boolean"
          }
        },
        "required": [
          "face_recognition_threshold",
          "recognition_cooldown_seconds",
          "attendance_min_interval_minutes",
          "camera_fps",
          "max_faces",
          "camera_id",
          "save_events",
          "model_version"
        ],
        "title": "ConfigOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
...
      "PrivacyOut": {
        "description": "Garantías fijas de privacidad. No son configurables: el sistema no tiene\nningún código que guarde imágenes o vídeo, ni envía datos fuera del equipo.",
        "properties": {
          "embeddings_encrypted": {
            "default": true,
            "title": "Embeddings Encrypted",
            "type": "boolean"
          },
          "local_processing": {
            "default": true,
            "title": "Local Processing",
            "type": "boolean"
          },
          "save_images": {
            "default": false,
            "title": "Save Images",
            "type": "boolean"
          },
          "save_video": {
            "default": false,
            "title": "Save Video",
            "type": "boolean"
          }
        },
        "title": "PrivacyOut",
        "type": "object"
      },
...
      "RuntimeConfig": {
        "additionalProperties": false,
        "properties": {
          "attendance_min_interval_minutes": {
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Attendance Min Interval Minutes",
            "type": "integer"
          },
          "camera_fps": {
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Camera Fps",
            "type": "integer"
          },
          "camera_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.\\-]+$",
            "title": "Camera Id",
            "type": "string"
          },
          "face_recognition_threshold": {
            "maximum": 1.0,
            "minimum": 0.2,
            "title": "Face Recognition Threshold",
            "type": "number"
          },
          "max_faces": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Faces",
            "type": "integer"
          },
          "recognition_cooldown_seconds": {
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Recognition Cooldown Seconds",
            "type": "integer"
          },
          "save_events": {
            "title": "Save Events",
            "type": "boolean"
          }
        },
        "required": [
          "face_recognition_threshold",
          "recognition_cooldown_seconds",
          "attendance_min_interval_minutes",
          "camera_fps",
          "max_faces",
          "camera_id",
          "save_events"
        ],
        "title": "RuntimeConfig",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "last_login_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "ADMIN",
          "OPERATOR"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Cookie HttpOnly que entrega POST /api/auth/login.",
        "in": "cookie",
        "name": "facetrack_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nReconocimiento facial, asistencia y estadísticas con **procesamiento local**.\n\n**Autenticación.** `POST /api/auth/login` entrega una cookie de sesión HttpOnly;\nel navegador la envía sola en las demás llamadas. Todas las rutas la requieren\nsalvo `GET /api/health` y el login. Las marcadas *Solo ADMIN* exigen ese rol.\n\n**Errores.** Siempre `{\"detail\": str, \"code\": str}`; los 5xx agregan\n`request_id` (también en la cabecera `X-Request-ID` de toda respuesta), que\npermite encontrar el detalle en los logs.\n\n**Privacidad.** Las imágenes se procesan en memoria y se descartan; ninguna\nrespuesta incluye embeddings.\n",
    "title": "FaceTrack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/asistencia": {
      "get": {
        "description": "Filtros: `person_id`, `type` (ENTRY/EXIT), y `date` o `from`/`to` (días locales, incluidos).",
        "operationId": "list_attendance_api_asistencia_get",
        "parameters": [
          {
            "in": "query",
            "name": "person_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AttendanceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "description": "Un día (hora local)",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Un día (hora local)",
              "title": "Date"
            }
          },
          {
            "description": "Desde (incluido, hora local)",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Desde (incluido, hora local)",
              "title": "From"
            }
          },
          {
            "description": "Hasta (incluido, hora local)",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hasta (incluido, hora local)",
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendancePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Attendance",
        "tags": [
          "asistencia"
        ]
      }
    },
    "/api/asistencia/export.csv": {
      "get": {
        "description": "Mismos filtros que el listado, sin paginar. Fechas y horas en hora local.",
        "operationId": "export_attendance_csv_api_asistencia_export_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "person_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Person Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AttendanceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "description": "Un día (hora local)",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Un día (hora local)",
              "title": "Date"
            }
          },
          {
            "description": "Desde (incluido, hora local)",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Desde (incluido, hora local)",
              "title": "From"
            }
          },
          {
            "description": "Hasta (incluido, hora local)",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hasta (incluido, hora local)",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Export Attendance Csv",
        "tags": [
          "asistencia"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Valida usuario y contraseña y abre una sesión en una cookie HttpOnly\n(no accesible desde JavaScript). Mensaje genérico ante cualquier error de credenciales.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Demasiados intentos"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Revoca el token actual (aunque alguien lo haya copiado, deja de servir) y borra la cookie.",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/configuracion": {
      "get": {
        "operationId": "get_config_api_configuracion_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Config",
        "tags": [
          "configuracion"
        ]
      },
      "put": {
        "description": "Reemplaza la configuración completa. Se guarda en la base y se aplica\nde inmediato (threshold, cooldown, máximo de rostros, etc.).",
        "operationId": "update_config_api_configuracion_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Update Config",
        "tags": [
          "configuracion"
        ]
      }
    },
...
  },
...
}